requirement.txt
django==4.1.7
djangorestframework==3.14.0
mysqlclient
pillow==9.4.0
django-jazzmin==2.6.0
django-filter==22.1